refactor(KeySequence): move KeySequence to common
