+++ src/lib/common/KeySequence.cpp
