+++ src/lib/gui/Hotkey.h
@@ -13,7 +13,7 @@
 #include <QTextStream>
 
 #include "Action.h"
-#include "KeySequence.h"
+#include "common/KeySequence.h"
 
 class HotkeyDialog;
 class ServerConfigDialog;
