+++ src/unittests/common/CMakeLists.txt
@@ -21,3 +21,10 @@ create_test(
   SOURCE I18NTests.cpp
   WORKING_DIRECTORY "${CMAKE_BINARY_DIR}/src/lib/common"
 )
+
+create_test(
+  NAME KeySequenceTests
+  DEPENDS common
+  SOURCE KeySequenceTests.cpp
+  WORKING_DIRECTORY "${CMAKE_BINARY_DIR}/src/lib/common"
+)
